# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema POS",
        "description": "Documentación de la API del Sistema POS",
        "contact": {
            "email": "admin@sistema.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/compras": {
            "get": {
                "tags": [
                    "Compras"
                ],
                "summary": "Obtener lista de compras",
                "operationId": "aec9e7fb057675eed94ae46fbd5141d2",
                "parameters": [
                    {
                        "name": "proveedor_id",
                        "in": "query",
                        "description": "Filtrar por proveedor",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "PENDIENTE",
                                "COMPLETADA",
                                "ANULADA"
                            ]
                        }
                    },
                    {
                        "name": "fecha_inicio",
                        "in": "query",
                        "description": "Fecha inicial para filtrar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_fin",
                        "in": "query",
                        "description": "Fecha final para filtrar",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de compras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "numero": {
                                                        "type": "string"
                                                    },
                                                    "proveedor": {
                                                        "type": "object"
                                                    },
                                                    "total": {
                                                        "type": "number"
                                                    },
                                                    "estado": {
                                                        "type": "string"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compras"
                ],
                "summary": "Crear una nueva compra",
                "operationId": "7d7b7f238f2bb0700c471ad967038783",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "proveedor_id",
                                    "tipo_compra",
                                    "detalles"
                                ],
                                "properties": {
                                    "proveedor_id": {
                                        "type": "integer"
                                    },
                                    "tipo_compra": {
                                        "type": "string",
                                        "enum": [
                                            "CONTADO",
                                            "CREDITO"
                                        ]
                                    },
                                    "fecha_vencimiento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "factura_proveedor": {
                                        "type": "string"
                                    },
                                    "nota": {
                                        "type": "string"
                                    },
                                    "detalles": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "producto_id",
                                                "unidad_medida_id",
                                                "cantidad",
                                                "precio"
                                            ],
                                            "properties": {
                                                "producto_id": {
                                                    "type": "integer"
                                                },
                                                "unidad_medida_id": {
                                                    "type": "integer"
                                                },
                                                "cantidad": {
                                                    "type": "integer"
                                                },
                                                "precio": {
                                                    "type": "number"
                                                },
                                                "descuento": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compra creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "numero": {
                                            "type": "string"
                                        },
                                        "proveedor": {
                                            "type": "object"
                                        },
                                        "total": {
                                            "type": "number"
                                        },
                                        "estado": {
                                            "type": "string"
                                        },
                                        "detalles": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}